abstract data processing into a util file
